Add shouldSeed variable to prevent erasing existing tables; test successful

+++ pulse-core-psql/create-dashboards.js
@@ -1,20 +1,22 @@
-const createDashboards = async sequelize => {
+const createDashboards = async (sequelize, shouldSeed) => {
   const Dashboard = await sequelize.import('dashboard', require('./models/dashboard'))
   Dashboard.belongsTo(Dashboard)
   Dashboard.hasMany(Dashboard, { onDelete: 'cascade' })
 
-  await Dashboard.sync({ force: true })
+  if (shouldSeed) {
+    await Dashboard.sync({ force: true })
 
-  await Dashboard.bulkCreate([
-    { name: 'Provider Targeted Accounts' },
-    { name: 'Payer' },
-  ])
+    await Dashboard.bulkCreate([
+      { name: 'Provider Targeted Accounts' },
+      { name: 'Payer' },
+    ])
 
-  const parentDashes = await Dashboard.findAll()
-  for (childDash of ['Management', 'Accounts']) {
-    for (parentDash of parentDashes) {
-      const createdChildDash = await Dashboard.create({ name: childDash })
-      await parentDash.addDashboard(createdChildDash)
+    const parentDashes = await Dashboard.findAll()
+    for (childDash of ['Management', 'Accounts']) {
+      for (parentDash of parentDashes) {
+        const createdChildDash = await Dashboard.create({ name: childDash })
+        await parentDash.addDashboard(createdChildDash)
+      }
     }
   }
